# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Investor Buddy API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  email: string;
}

export interface Stock {
  ticker: string;
  name: string;
  sector: string;
  price: number;
  peRatio: number;
  debtToEquity: number;
  epsGrowth: number;
  dividendYield: number;
  roe: number;
  revenueGrowth: number;
  profitMargin: number;
  currentRatio: number;
  marketCapB: number;
  beta: number;
  updatedAt: string;
}

export type CriterionOperator =
  (typeof CriterionOperator)[keyof typeof CriterionOperator];

export const CriterionOperator = {
  ">": ">",
  "<": "<",
  ">=": ">=",
  "<=": "<=",
  between: "between",
} as const;

export interface Criterion {
  metric: string;
  operator: CriterionOperator;
  value: number;
  /** @nullable */
  value2?: number | null;
}

export interface CriteriaSet {
  id: string;
  name: string;
  criteria: Criterion[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateCriteriaSetRequest {
  /** @nullable */
  id?: string | null;
  name: string;
  criteria: Criterion[];
}

export interface ScreenRequest {
  criteria: Criterion[];
}

export interface CriterionResult {
  metric: string;
  passed: boolean;
  value: number;
}

export interface ScreenResult {
  ticker: string;
  name: string;
  sector: string;
  score: number;
  price: number;
  peRatio: number;
  debtToEquity: number;
  epsGrowth: number;
  dividendYield: number;
  roe: number;
  revenueGrowth: number;
  profitMargin: number;
  currentRatio: number;
  marketCapB: number;
  beta: number;
  criteriaResults: CriterionResult[];
}

export interface WatchlistItem {
  ticker: string;
  name: string;
  sector: string;
  price: number;
  peRatio?: number | null;
  dividendYield?: number | null;
  addedAt: string;
}

export interface AnalyzeRequest {
  ticker: string;
  stock: Stock;
  criteria: Criterion[];
  score: number;
}

export interface AnalyzeResponse {
  explanation: string;
}

export interface DashboardSummary {
  totalStocks: number;
  totalCriteriaSets: number;
  watchlistCount: number;
  sectorsCount: number;
  avgPeRatio: number;
  avgEpsGrowth: number;
}

export interface SectorCount {
  sector: string;
  count: number;
}

export interface StockSearchResult {
  ticker: string;
  name: string;
  type: string;
  region: string;
  inDatabase: boolean;
}

export type RefreshAllStocks202 = {
  message: string;
};