# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Investor Buddy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks:
    get:
      operationId: getStocks
      tags: [stocks]
      summary: Get all stocks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stock"
  /stocks/search/{query}:
    get:
      operationId: searchStocks
      tags: [stocks]
      summary: Search for stocks by name or ticker via Alpha Vantage
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/add/{ticker}:
    post:
      operationId: addStock
      tags: [stocks]
      summary: Add a stock by fetching from Alpha Vantage
      security:
        - bearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/refresh/{ticker}:
    post:
      operationId: refreshStock
      tags: [stocks]
      summary: Refresh a single stock's data from Finnhub
      security:
        - bearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /stocks/refresh-all:
    post:
      operationId: refreshAllStocks
      tags: [stocks]
      summary: Refresh prices for all stocks via Finnhub (background job)
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Refresh started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /stocks/{ticker}:
    get:
      operationId: getStock
      tags: [stocks]
      summary: Get a stock by ticker
      security:
        - bearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteStock
      tags: [stocks]
      summary: Delete a stock from the database
      security:
        - bearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Stock:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        sector:
          type: string
        price:
          type: number
        peRatio:
          type: number
        debtToEquity:
          type: number
        epsGrowth:
          type: number
        dividendYield:
          type: number
        roe:
          type: number
        revenueGrowth:
          type: number
        profitMargin:
          type: number
        currentRatio:
          type: number
        marketCapB:
          type: number
        beta:
          type: number
        updatedAt:
          type: string
      required:
        - ticker
        - name
        - sector
        - price
        - peRatio
        - debtToEquity
        - epsGrowth
        - dividendYield
        - roe
        - revenueGrowth
        - profitMargin
        - currentRatio
        - marketCapB
        - beta
        - updatedAt
...
    StockSearchResult:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        type:
          type: string
        region:
          type: string
        inDatabase:
          type: boolean
      required:
        - ticker
        - name
        - type
        - region
        - inDatabase